Move tests from gost89.c into test_gost89.c.
